core: Fix support for overriding PRODUCT_DEVICE and PRODUCT_NAME
* Previous commit "core: Add support for overriding build.prop settings"
is broken that it does not provide the correct ability of overriding
PRODUCT_DEVICE and PRODUCT_NAME, that we had on android R.

* Only override PRODUCT_SYSTEM_NAME and PRODUCT_SYSTEM_DEVICE when they
are previously not equal to the vendor one, because they can be assgined
with "generic" and "mainline" when building generic system.

Test:

Set
```
PRODUCT_BUILD_PROP_OVERRIDES += \
    PRODUCT_DEVICE=OnePlus8T \
    PRODUCT_NAME=OnePlus8T_EEA \
```
build and check props.

Change-Id: I3f48e638b94b9f055352cd534acadeae35fbe91a
Signed-off-by: LibXZR <i@xzr.moe>